block: expose inline encryption support through a request queue flag
Define a request queue flag QUEUE_FLAG_INLINECRYPT to indicate that the
underlying block device supports inline encryption, and enable it in the
ufs-qcom driver (via an intermediate flag on the Scsi_Host) when the ICE
support is compiled in and the ICE hardware has been detected.
Filesystems will be able to use this flag to decide whether they should
allow the use of inline encryption.

(Of course, this comes with the important caveat that there isn't yet a
proper API to actually *use* inline encryption, besides PFK and
dm-req-crypt, but that's arguably a separate issue.  Also, note that we
can't simply make PFK depend on SCSI_UFS_QCOM_ICE because the
dependencies already go in the opposite direction: SCSI_UFS_QCOM_ICE
depends on CRYPTO_DEV_QCOM_ICE which depends on PFK, plus it really
would be SCSI_UFS_QCOM_ICE *or* MMC_SDHCI_MSM_ICE that is needed,
depending on the hardware.)

Change-Id: Ie1bb4cf9ece21e1ba7e9b4298031df2a2fe7e94b
Bug: 33042316
Signed-off-by: Eric Biggers <ebiggers@google.com>